move the calculator files around
