+++ examples/flutter_gallery/lib/demo/calculator_logic.dart
