+++ examples/flutter_gallery/lib/gallery/app.dart
@@ -5,7 +5,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart' show timeDilation;
 
-import '../calculator/interface.dart';
 import '../demo/all.dart';
 import 'home.dart';
