+++ examples/flutter_gallery/lib/gallery/home.dart
@@ -5,7 +5,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 
-import '../calculator/interface.dart';
 import '../demo/all.dart';
 import 'drawer.dart';
 import 'header.dart';
